requirements.txt
pandas==2.2.3
numpy==2.1.1
scikit-learn==1.5.2
Flask==3.0.3
flask-cors==4.0.1
streamlit==1.38.0
SQLAlchemy==2.0.32
psycopg2-binary==2.9.9
python-dotenv==1.0.1
joblib==1.4.2
pydantic==2.9.2
pytest==8.3.3
requests==2.32.3
plotly==5.24.0
orjson==3.10.7
